# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pulsar_data_aggregation_API",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/data-aggregation-plan-contract": {
            "get": {
                "tags": [
                    "Plan contract"
                ],
                "operationId": "c89900359ef20cffaef4627a760c0193",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "Plan contract"
                ],
                "operationId": "1aeecf204463389259dfaede634579c1",
                "requestBody": {
                    "request": "Create Plan contract",
                    "description": "Create Plan contract Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "company_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "date_contract_end": {
                                        "description": "0",
                                        "type": "date",
                                        "example": "2022-01-01"
                                    },
                                    "sum": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "company_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "year": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "short_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "company_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        },
                                        "date_contract_end": {
                                            "description": "0",
                                            "type": "date",
                                            "example": "2022-01-01"
                                        },
                                        "sum": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "company_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "year": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "short_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/data-aggregation-plan-contract/get-by-id/{id}": {
            "get": {
                "tags": [
                    "Plan contract"
                ],
                "operationId": "12024582c14bdce44eedc22f49e47ec9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/data-aggregation-plan-contract/get-by-name/{company_name}": {
            "get": {
                "tags": [
                    "Plan contract"
                ],
                "operationId": "354f224cd5199c718fbaea56aff389a9",
                "parameters": [
                    {
                        "name": "company_name",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/data-aggregation-plan-contract/{id}": {
            "delete": {
                "tags": [
                    "Plan contract"
                ],
                "operationId": "df4089483fe15ea421589273c2c00198",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Plan contract"
                ],
                "operationId": "499248874d9f25c1054314c34439c2c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Update Plan contract",
                    "description": "Update Plan contract Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "company_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "date_contract_end": {
                                        "description": "0",
                                        "type": "date",
                                        "example": "2022-01-01"
                                    },
                                    "sum": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "company_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "year": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "short_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "company_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        },
                                        "date_contract_end": {
                                            "description": "0",
                                            "type": "date",
                                            "example": "2022-01-01"
                                        },
                                        "sum": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "company_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "year": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "short_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                },
                                                "field2": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field2 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}